# docs/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: Hotel Search API
  description: REST API for hotel CRUD and natural-language search (Lemax take-home
    PoC). Write operations may require the X-Api-Key header when ApiKey:WriteKey is
    configured.
  version: v1
servers:
- url: http://localhost:5103
  description: Local dotnet run
- url: http://localhost:8080
  description: Docker Compose
paths:
  /api/hotels:
    post:
      tags:
      - Hotels
      operationId: CreateHotel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHotelRequest'
            example:
              name: Grand Hotel Zagreb
              price: 120
              latitude: 45.815
              longitude: 15.9819
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - X-Api-Key: []
    get:
      tags:
      - Hotels
      operationId: ListHotels
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelDto'
  /api/hotels/{id}:
    get:
      tags:
      - Hotels
      operationId: GetHotelById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
      - Hotels
      operationId: UpdateHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHotelRequest'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
      - Hotels
      operationId: DeleteHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/hotels/search:
    post:
      tags:
      - Search
      operationId: SearchHotels
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchHotelsRequestDto'
            example:
              prompt: near 45.8150, 15.9819 under 200
              page: 1
              pageSize: 10
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseOfSearchHotelItemDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    CreateHotelRequest:
      type: object
      properties:
        name:
          type: string
        price:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
          - number
          - string
          format: double
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
    HotelDto:
      required:
      - id
      - name
      - price
      - latitude
      - longitude
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
          - number
          - string
          format: double
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
    PagedResponseOfSearchHotelItemDto:
      required:
      - items
      - page
      - pageSize
      - totalCount
      - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchHotelItemDto'
        page:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
    ProblemDetails:
      type: object
      properties:
        type:
          type:
          - 'null'
          - string
        title:
          type:
          - 'null'
          - string
        status:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - 'null'
          - integer
          - string
          format: int32
        detail:
          type:
          - 'null'
          - string
        instance:
          type:
          - 'null'
          - string
    SearchHotelItemDto:
      required:
      - id
      - name
      - price
      - distanceKm
      - rankingScore
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
          - number
          - string
          format: double
        distanceKm:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
        rankingScore:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
    SearchHotelsRequestDto:
      type: object
      properties:
        prompt:
          type: string
        page:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
    UpdateHotelRequest:
      type: object
      properties:
        name:
          type: string
        price:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
          - number
          - string
          format: double
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
          type:
          - number
          - string
          format: double
  securitySchemes:
    X-Api-Key:
      type: apiKey
      description: Optional API key for POST/PUT/DELETE when ApiKey:WriteKey is configured
        on the server.
      name: X-Api-Key
      in: header
tags:
- name: Hotels
- name: Search